Add Check and CompatibilityType interfaces: Introduce Check interface for condition evaluation and CompatibilityType for compatibility status representation, enhancing type safety and modularity in checks.
